fix build java

+++ langs/build.java.sh
@@ -4,7 +4,9 @@ if [[ -f $gitlab_project_dir/build.gradle ]]; then
     echo_time_step "java build [gradle]..."
     gradle -q
 else
-    echo_time_step "java build [maven]..."
-    docker run -i --rm -v "$gitlab_project_dir":/usr/src/mymaven -w /usr/src/mymaven maven:3.3-jdk-8 mvn clean install
+    if [[ "${ENV_BUILD_JAVA_USE_MVN:-0}" = 1 ]]; then 
+        echo_time_step "java build [maven]..."
+        docker run -i --rm -v "$gitlab_project_dir":/usr/src/mymaven -w /usr/src/mymaven maven:3.3-jdk-8 mvn clean install
+    fi
     # mvn clean install
 fi
